Updated user query code - thx to RaymondSchnyder
So I am unsure if this works because I have not run the code myself, though it should work. I`ll do final changes in a bit. Now to query for a user, all you need to do is get some first part of the user`s username, so for example if the user`s name is ¨bob johnson¨, you should be able to type in bob, and as long as there is no other user with the same first three characters, it should be able to find him.